+++ test.html
@@ -77,7 +77,7 @@
 
     <div class="test-section">
         <h3>要素の追加・削除テスト</h3>
-        <button class="test-button" onclick="addElement()">うんち</button>
+        <button class="test-button" onclick="addElement()">クラスを変更fdfa</button>
         <button class="test-button" onclick="removeElement()">要素</button>
         <div id="dynamic-content"></div>
     </div>
@@ -95,7 +95,7 @@
     <div class="test-section">
         <h3>テキスト変更テスト</h3>
         <input type="text" class="test-input" id="text-input" value="変更前のテキスト" onchange="updateText(this.value)">
-        <div id="text-test" style="color: red;font-weight: bold;">>ふぁｆ</div>
+        <div id="text-test" style="color: red;font-weight: bold;">>ふぁｆfdfds</div>
     </div>
 
     <script>
@@ -236,6 +236,9 @@
 <script src="chrome-extension://gppongmhjkpfnbhagpmjfkannfbllamg/js/js.js"></script>
 <script src="chrome-extension://gppongmhjkpfnbhagpmjfkannfbllamg/js/js.js"></script>
 <script src="chrome-extension://gppongmhjkpfnbhagpmjfkannfbllamg/js/dom.js"></script>
+<script src="chrome-extension://gppongmhjkpfnbhagpmjfkannfbllamg/js/js.js"></script>
+<script src="chrome-extension://gppongmhjkpfnbhagpmjfkannfbllamg/js/dom.js"></script>
+<script src="chrome-extension://gppongmhjkpfnbhagpmjfkannfbllamg/js/js.js"></script>
 <script src="chrome-extension://gppongmhjkpfnbhagpmjfkannfbllamg/js/js.js"></script>
 <script src="chrome-extension://gppongmhjkpfnbhagpmjfkannfbllamg/js/dom.js"></script>
 <script src="chrome-extension://gppongmhjkpfnbhagpmjfkannfbllamg/js/js.js"></script>
