--- vscode-extention/src/test/extension.test.ts
@@ -1,12 +1,0 @@
-// テストファイルを一時的に無効化
-// import * as assert from 'assert';
-// import * as vscode from 'vscode';
-
-// suite('Extension Test Suite', () => {
-// 	vscode.window.showInformationMessage('Start all tests.');
-
-// 	test('Sample test', () => {
-// 		assert.strictEqual(-1, [1, 2, 3].indexOf(5));
-// 		assert.strictEqual(-1, [1, 2, 3].indexOf(0));
-// 	});
-// });
